pyproject.toml
[tool.poetry]
name = "ankichat"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests-html = "^0.10.0"
aiofiles = "23.2.1"
aiogram = "3.7.0"
aiohttp = "3.9.5"
aiosignal = "1.3.1"
aiosqlite = "0.20.0"
annotated-types = "0.7.0"
async-timeout = "4.0.3"
asyncpg = "0.29.0"
attrs = "23.2.0"
build = "1.2.1"
cachecontrol = "0.14.0"
certifi = "2024.6.2"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
crashtest = "0.4.1"
cryptography = "42.0.8"
distlib = "0.3.8"
dulwich = "0.21.7"
fastjsonschema = "2.20.0"
filelock = "3.15.4"
frozenlist = "1.4.1"
greenlet = "3.0.3"
idna = "3.7"
importlib-metadata = "7.2.1"
installer = "0.7.0"
jaraco-classes = "3.4.0"
jeepney = "0.8.0"
keyring = "24.3.1"
magic-filter = "1.0.12"
more-itertools = "10.3.0"
msgpack = "1.0.8"
multidict = "6.0.5"
packaging = "24.1"
pexpect = "4.9.0"
pkginfo = "1.11.1"
platformdirs = "4.2.2"
poetry = "1.8.3"
poetry-core = "1.9.0"
poetry-plugin-export = "1.8.0"
psycopg2 = "2.9.9"
ptyprocess = "0.7.0"
pycparser = "2.22"
pydantic = "2.7.4"
pydantic-core = "2.18.4"
pyproject-hooks = "1.1.0"
python-dotenv = "1.0.1"
rapidfuzz = "3.9.3"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
secretstorage = "3.3.3"
shellingham = "1.5.4"
sqlalchemy = "2.0.30"
tomlkit = "0.12.5"
trove-classifiers = "2024.5.22"
typing-extensions = "4.12.2"
urllib3 = "2.2.2"
virtualenv = "20.26.3"
yarl = "1.9.4"
zipp = "3.19.2"
pydub = "^0.25.1"
humanize = "^4.10.0"
i18n = "^0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
